backend/app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.8
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
inference-sdk==0.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.15.1
six==1.17.0
supervision==0.25.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
